Rearrange DMA ISR to inline regenerate_waveform (992->968)
This change arranges the ISR so that it tailcalls regenerate_waveform
exactly once per invocation, which in turn frees the compiler to inline
it.

In the unlikely event where the ISR triggers because both HTIF and TCIF
are set, this will cause an immediate re-invocation of the ISR, which is
a small performance hit -- but a very unlikely one.