Fix integer overflow in quiet time calculation.
In the original, the quiet-time counter would overflow every 8 hours 17
minutes, resulting in a brief (8ms, 333 cycles) and very quiet burst of
carrier. I reproduced this in the initial rewrite by using fetch_add,
which is a wrapping add.

Here I've switched it to saturate, which seems to make more sense.