Convert the wavetable to u16s.
The DAC samples are 12 bits, right-aligned in a 32-bit register. This
register does not support halfword accesses. However, the DMA controller
on the F0 is happy to zero-extend values when PSIZE != MSIZE.

So, this change switches the wave in memory to use 16-bit samples, and
has the DMA controller extend on transfer. This recovers the memory I
spent on double-buffering the waveform.