Compute sine coefficients at build time.
So it turns out, the hand-constructed sine coefficients in the original
system were slightly wrong. I know, shocking.

This generates the sine table at build time, providing a full 16 bits of
precision. Not that it probably matters in practice -- the errors in the
table will have been introducing some noise at 500-1000 kHz, which
probably doesn't make it through the amp, much less the transducers. But
hey, now this is easier to maintain!