Initialize DAC at startup to avoid negative pulse.
By leaving the data register at its reset value of zero, the original
system (and versions of this firmware up until now) would drive the
output toward 0V until the first waveform started being transmitted.
That's just unhygenic, so, it is now initialized to the zero midpoint
immediately.

The DAC is also now initialized _before_ its pin is set to analog mode,
so it remains in high-impedance during initialization to avoid glitches.